An inelegant brute-forcey solution
It's the right answer.  The puzzle says so.  But I haven't really proven
it.  Not even to myself.  I mean, the bigger the square the less likely
the power will be high because on average numbers are negative.  But
definitely not rigorous.  Especially since the hash we're using to
generate powers is super primitive and seems like it could leave all
kinds of surprising and gross artifacts.

A more efficent solution would have had me not recounting squares.

Even better starting from the bottom right and heading up.

A probably more efficient solution still would have had power-of-two
sized squares stored and sum the ones that fit.  But so many fenceposts
to check...

I should probably do one of those...