:facepalm: Problem was Copy-paste error on input
So the only problem was that I was missing the top line of the input
this was clear to me only after running the simulation and watching it a
few times and then wondering why there was one track on the right side
that went up into nowhere.  Doh!

I commented out the part that makes this pretty and it just gives the
answer but it's easy enough to add back in and kind of fun to watch the
little trains drive around.