List removal is ghetto but solution correct
Left in the drawing code.  Probably fun to watch all the trains
crash.  It would actually be sort of a cool game to guess which
train(s) may survive.  Speaking of plural trains surviving... I don't
handle that.  It'll just keep looping indefinitely which is fine when
you're watching the visualization but less cool when you're expecting
an answer.