Day 14
Part A works fine.  It's easy.  My particular input for part B exhausts
my memory though.  For several other inputs I've found it works fine.
The problem is that there is a pseudo-random self-refernetial sequence
and you have to find the first instance of a pattern.  Since the pattern
is self-referential, you have to keep at least part of it around.

I have a naive-ish solution in Csharp and C++ in which I keep exactly 1
copy of the sequence around.  It could be made more optimal if instead
of keeping all values, I just kept the parts potentially needed again
which is on the order of 1 in 4 of the values.  (Since all paths will
eventually (and quickly!) converge on one path through the sequence).
This still doesn't convince me that I'll have enough memory though and
so I haven't done the work to optimize that step.  A sequence of 100
million entries contains only about 80% of all possible 6 digit inputs.
Increasing to 1 billion increase the coverage by only a few percent
which implies that getting to 10 billion 100 billion entries still
may not get me there.

So either the luck of your input really matters (I doubt) or expect I'm
missing some clever realization about the nature of the sequence and
the need to keep a large subset of it in memory.

Anyway, failure #1.  May revisit later.