2020 Day 6
This would have been done in 5 minutes in python.  But part B took me
quite a bit longer in Rust as I struggled to make the HashSet
multi-intersection out of a fold/reduce.  I tried some other ways but
I sure like an elegant functional-style solution.

I really dislike that HashSet::intersection doesn't return a Hashset,
but rather an "Intersection" which is an iterator.  What's more, it
iterates over references to the underlying data meaning that you can't
easily nest them because the type changes to a reference.

Presumably this is done for efficiency/flexibility in some other case,
but golly it tripped me up trying to make the semi-deeply-nested types
line up just as expected.  The internet is filled with other casualties
of this decision.