Day 7
Fun with recursion, also building and navigating simple graphs, also,
spending another few hours fighting the rust regex library which is
pretty damned barebones.

No backreferences?  No lookbehinds?  No capturing multiple groups of
groups?  Lame.  I had to actually break my regex up into two separate
regexes to solve the parsing for this problem.

And not to keep being a big downer on Rust, because it is a lovely
language, but the immutability of the keys on HashMap are a real bummer.
I used BTreeMap here as a replacement which worked fine for thise.