Day 8
Moderately enjoyable.  A bit strange that NOP has an "argument", but ok!

It is clear I need to learn more about what it means to clone a vector.
It makes sense that to clone a vector (in the sense of creating a deep
copy) would require that individual elements be Cloneable.  But I'm not
sure why what I get back from a clone() on a vector is a reference and
not a moved value like I apparently get with to_vec().

In any case, the program works fine.