Day 10 - A worthy foe
I think that part A took about 10 minutes.  But Part B caused me to
scratch my head for a couple hours on how to actually arrive at and
implement a solution.

My first few solutions failed.

I was surprised to find that what seemed like a combinatorics problem
was answered not through multiplication but a sort of very limited
recursive addition.

I also learned some fun new Rust tricks.  The "Entry" api on associative
arrays is the replacement for my beloved setdefault in python.

Curious to see if anyone implemented this one in a fundamentally
different way than I did.