Day 11 - Fun with iterators
I decided to stretch myself on this one a little bit and build the
solution to part A by creating an iterator.  I figured that it was
likely that part B would change the chairs I would need to consider and
so having a flexible system in place would be useful.  Also, iterators
are fantastic and I've come to really like the features you can derive
from them in Rust.

The iterators are messy.  Nested ifs and loops.  Pretty gross looking.
I expect I could clean them up so that their logic would be much more
clear but for now I'm tuckered.

Another new bit here is tests.  I had several off-by-ones errors in my
tests.  One even produced a different evaluation of the chairs, but that
happened to have the same number of steps and the same answer for the
example input.  Building tests around these complex functions worked
well except that with some many binaries in a project, the tests output
is verbose.