Day 14 - Fun with Bitwise Operations
I'm a bit bummed that my program doesn't handle the example input for
part a in part b.  Doing so would require too many billions of operations.  There may be a slicker way to handle the calculation when the number of Xs in the mask in part B approaches high values--this is an exponential increase in expense.