Day 17 - Conway good, Part B Lame
Just adding another dimension in part B was kinda lame.  I guess maybe
the lesson is that if there were 3 dimensions, I maybe should have just
set it up for N dimensions in the first place so that it became just a
constant change?  Except the X, Y dimensions were already special-cased
on the input.  And for outputting.