Expect calculators to solve both parts
Updated the framework so that calculators are expected to calculate
solutions to both parts A and B.  I had wanted to have the Solution
trait have a default implementation that failed with unimplemented but
couldn't quickly get that working.  I don't mind that so much though
because the compiler reminds you to implement something.