Implemnet 2022/day03
So the algorithm here was super easy.  Just use some sets and some set
operations.  No problem.

The part B was actually pretty costly though.  The way Rust has set
intersections and unions implemented, you can't just nicely chain them
and everything has to be a reference.  If I stopped and thought about
*why* that is, there is probably a good reason for efficiency but may it
makes the implementation look crappy compared to just a series of '&'
and '|' operations or chained intersection/union ops.

It's ultimately not a huge thing, but it's one of the more negative Rust
experiences I've had recently.