Implemnet 2022 day05
So I kind of fell out of love with the rust regex library here.  I was
not able to set up groups of repeated captures like I can in python and
that made it pretty annoying.  Parsing became a two step process due to
the need to know that the whole line matched and figure out its length.

It was a big enough pain that I switched to just grabbing the letter out
by index which is a bit more error prone and fragile.