Implement 2022 day06
Nice and easy.  I solve part A in a very specific simple way with my
son rather expecting that part B would want a more general solution.  It
did and so I implemented that with sets rather than O(N^2) comparisons.

Depending on the size of N, the specifc solution in A is probaly more
efficient.  For a large N, keeping the previous set around and removing
the "head" and inserting the "tail" is more efficient still.  But for
these sizes, that optimization is quite irrelevant.