Implement 2022 day 08
Yup.  I worked on day 7 for a while but it wore me down.  The parsing
was pretty tedious.  The other problem is that the directory structure
is self-referencing and a thing I wanted to mutate which caused all
kinds of issues with the Complainer.  Stuff I "knew" was safe but I
couldn't be convincing on.  Lots of ways around it but none of them were
fun at the time so I worked on day 08 instead.

Day 08 was pretty easy.  I busted out the Grid library which made life
nicer then I fumbled a bunch with reverse numeric ranges.